requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
pillow~=8.1.0
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
tqdm>=4.41.0

pyrealsense2~=2.41.0.2666
torch==1.6.0
torchvision==0.7.0



yaml~=0.2.5
sympy~=1.7.1
requests~=2.25.1
pyserial~=3.5